+++ src/main/kotlin/uk/co/dmatthews/khip8/memory/TimerRegister.kt
@@ -4,7 +4,7 @@ import kotlinx.coroutines.delay
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
-class TimerRegister(var value: UByte = 0.toUByte(), private var halt: Boolean = false) {
+class TimerRegister(var value: UByte = 0u, private var halt: Boolean = false) {
 
     fun halt() {
         halt = true
@@ -25,6 +25,6 @@ class TimerRegister(var value: UByte = 0.toUByte(), private var halt: Boolean = 
         // 60 Hz, calculated as 1000 / 60 = 16.66666666666667 rounded up to 17. This is slightly inaccurate, clocking
         // in at 1020 millis every 60 ticks
         const val FREQUENCY_IN_MILLIS = 17L
-        private val ZERO = 0.toUByte()
+        private val ZERO = 0u
     }
 }
